Move PtProject class to projects folder
