+++ model/index.js
@@ -1,5 +1,4 @@
 const clients = require('restify-clients')
-const {getProject, getAllProjects} = require("./projects")
 const {getIterations, getIterationCycleTime} = require("./iterations")
 const {getMemberships} = require("./accounts")
 const {getAllTasks, deliverTask, undeliverTask} = require('./tasks')
@@ -8,8 +7,6 @@ const {common} = require('../lib/commands/common')
 const {normaliseFields} = require('../lib/adapters/normaliseFields')
 
 const model = {
-  getProject,
-  getAllProjects,
   getIterationCycleTime,
   getIterations,
   getMemberships,
@@ -70,30 +67,7 @@ class Model {
   }
 }
 
-class PtProject extends Model {
-  constructor(context) {
-    super(context)
-    this._baseProjectPath = this._baseApiPath
-  }
-
-  get _endpoints() {
-    return {
-      getProject: fields => this._appendFields(this._baseProjectPath, fields),
-      getAllProjects: fields => this._appendFields(`/services/v5/projects`, fields)
-    }
-  }
-
-  getProject(fields = []) {
-    return this._fetch('get', this._endpoints.getProject(fields))
-  }
-
-  getAllProjects(fields = []) {
-    return this._fetch('get', this._endpoints.getAllProjects(fields))
-  }
-}
-
 module.exports = {
   model,
-  Model,
-  PtProject
+  Model
 }
